make first 12-ball auto hopefully? untested...
