--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autoControl.java
@@ -1,225 +1,0 @@
-// Written primarily by Henry Rosenberg for AcaBots, FTC Team #24689
-// hi also sze ting
-// and Tsimur Havarko
-/*
--------------------- CONTROL SCHEME - CONTROLLER 1 --------------------
-Buttons:
-    A:
-    B:
-    X:
-    Y:
-
-D-Pad:
-    UP:
-    DOWN:
-    RIGHT:
-
-Triggers:
-    RT: Hold for Intake
-    LT: Hold for Reverse Intake
-    BOTH:
-
-Shoulder Buttons:
-    RB:
-    LB:
-
-Joysticks:
-    Right: Relative Chassis Rotation
-    Left: Absolute Chassis Strafe based on orientation when START button is pressed
-
--------------------- CONTROL SCHEME - CONTROLLER 2 --------------------
-Buttons:
-    A: Hold for Flywheel
-    B: Hold for Reverse Flywheel
-    Y:
-
-Triggers:
-    RT: Hold for Belts
-    LT: Hold for Reverse Belts
-
-Joysticks:
-    Right: Relative Chassis Rotation
-    Left: Absolute Chassis Strafe based on orientation when START button is pressed
-
- ---------------------------- START CONFIG ----------------------------
- Hanging Hooks: Open
- Arm Slide: Retracted
- Arm Pivot: Down, resting on bottom stop
- Claw Wrist: Folded left
- */
-
-package org.firstinspires.ftc.teamcode;
-import com.qualcomm.hardware.rev.Rev2mDistanceSensor;
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.hardware.CRServo;
-import com.qualcomm.robotcore.hardware.DistanceSensor;
-import com.qualcomm.robotcore.hardware.IMU;
-import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotorSimple;
-import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
-import com.qualcomm.robotcore.hardware.Servo;
-import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
-import org.firstinspires.ftc.robotcore.external.navigation.DistanceUnit;
-import org.firstinspires.ftc.robotcore.external.navigation.Pose2D;
-import org.firstinspires.ftc.robotcore.external.navigation.UnnormalizedAngleUnit;
-import java.util.Locale;
-
-@TeleOp
-public class autoControl extends LinearOpMode {
-    DcMotor frontLeftMotor; // 1
-    DcMotor backLeftMotor; // 0
-    DcMotor frontRightMotor; // 1 (expansion)
-    DcMotor backRightMotor; // 0 (expansion)
-    CRServo intakeLeft; // 1
-    CRServo intakeRight; // 0 (expansion)
-    CRServo beltLeft; // 0
-    CRServo beltRight; // 1 (expansion)
-    CRServo beltVertical; // 2
-    //    CRServo turret; // 2 (expansion)
-    DcMotor flywheelRotateMotor; // 2 (expansion)
-    DcMotor flywheelMotor; // 2
-    DcMotor flywheelIntake; // 3
-    Servo flywheelAngle; // 2 (expansion)
-    // CRServo clawIntake;
-    IMU imu;
-    DistanceSensor rightDistanceSensor;
-    DistanceSensor backDistanceSensor;
-    double angle;
-    GoBildaPinpointDriver odo;
-    double oldTime = 0;
-
-    private double calcLargestChange(double a, double b) {
-        // Return the value of the greatest absolute value of either a or b. Used for dual controller input
-        if (Math.abs(b) > Math.abs(a)) {
-            return b;
-        } else {
-            return a;
-        }
-    }
-
-    private int setSignFromReference(int newAbsoluteValue, int signReference) {
-        // Return the value of newAbsoluteValue with the + or - sign of signReference. Used for teleOp presets
-        if (signReference <= 0) {
-            return -newAbsoluteValue;
-        } else {
-            return newAbsoluteValue;
-        }
-    }
-
-    public void StopAll() {
-        intakeLeft.setPower(0);
-        intakeRight.setPower(0);
-        beltLeft.setPower(0);
-        beltRight.setPower(0);
-        beltVertical.setPower(0);
-        backLeftMotor.setPower(0);
-        frontLeftMotor.setPower(0);
-        backRightMotor.setPower(0);
-        frontRightMotor.setPower(0);
-        flywheelMotor.setPower(0);
-        flywheelIntake.setPower(0);
-    }
-
-    @Override
-    public void runOpMode() throws InterruptedException {
-        frontLeftMotor = hardwareMap.dcMotor.get("frontLeftMotor");
-        backLeftMotor = hardwareMap.dcMotor.get("backLeftMotor");
-        frontRightMotor = hardwareMap.dcMotor.get("frontRightMotor");
-        backRightMotor = hardwareMap.dcMotor.get("backRightMotor");
-
-        // Game Element Intake
-        intakeLeft = hardwareMap.get(CRServo.class, "intakeLeft");
-        intakeRight = hardwareMap.get(CRServo.class, "intakeRight");
-
-        // Belts to connect intake and flywheel
-        beltLeft = hardwareMap.get(CRServo.class, "beltLeft");
-        beltRight = hardwareMap.get(CRServo.class, "beltRight");
-        beltVertical = hardwareMap.get(CRServo.class, "beltVertical");
-
-        // Reverse some belts
-        intakeLeft.setDirection(CRServo.Direction.REVERSE);
-        beltLeft.setDirection(CRServo.Direction.REVERSE);
-
-        // Flywheel Motor and Rotation
-        flywheelRotateMotor = hardwareMap.dcMotor.get("rotatShot");
-        flywheelMotor = hardwareMap.dcMotor.get("flywheelMotor");
-        flywheelIntake = hardwareMap.dcMotor.get("flywheelIntake");
-        flywheelAngle = hardwareMap.get(Servo.class, "flywheelAngle");
-
-        // Reverse some of the drive motors depending on physical setup
-        frontLeftMotor.setDirection(DcMotorSimple.Direction.REVERSE);
-        backLeftMotor.setDirection(DcMotorSimple.Direction.REVERSE);
-
-        // Retrieve the IMU from the hardware map
-        imu = hardwareMap.get(IMU.class, "imu");
-        // Adjust the orientation parameters to match your robot <------------------------------------------------------- IMPORTANT
-        IMU.Parameters parameters = new IMU.Parameters(new RevHubOrientationOnRobot(
-                RevHubOrientationOnRobot.LogoFacingDirection.FORWARD,
-                RevHubOrientationOnRobot.UsbFacingDirection.RIGHT));
-        // Without this, the REV Hub's orientation is assumed to be logo up / USB forward
-        imu.initialize(parameters);
-
-        // Setup odometry params
-        odo = hardwareMap.get(GoBildaPinpointDriver.class,"odo");
-        odo.setOffsets(0, 0, DistanceUnit.INCH);
-        odo.setEncoderResolution(GoBildaPinpointDriver.GoBildaOdometryPods.goBILDA_4_BAR_POD);
-        odo.setEncoderDirections(GoBildaPinpointDriver.EncoderDirection.FORWARD, GoBildaPinpointDriver.EncoderDirection.FORWARD);
-
-        // Initialize odometry stuff when match starts
-        odo.resetPosAndIMU();
-        telemetry.addData("Status", "Initialized");
-        telemetry.addData("X offset", odo.getXOffset(DistanceUnit.INCH));
-        telemetry.addData("Y offset", odo.getYOffset(DistanceUnit.INCH));
-        telemetry.addData("Device Version Number:", odo.getDeviceVersion());
-        telemetry.addData("Heading Scalar", odo.getYawScalar());
-        telemetry.update();
-
-        // Retrieve the IMU from the hardware map
-        //imu = hardwareMap.get(IMU.class, "imu");
-        // Adjust the orientation parameters to match your robot <------------------------------------------------------- IMPORTANT
-        /*IMU.Parameters parameters = new IMU.Parameters(new RevHubOrientationOnRobot(
-                RevHubOrientationOnRobot.LogoFacingDirection.FORWARD,
-                RevHubOrientationOnRobot.UsbFacingDirection.RIGHT));*/
-        // Without this, the REV Hub's orientation is assumed to be logo up / USB forward
-        //imu.initialize(parameters);
-
-        // Chassis-mounted distance sensors
-//        rightDistanceSensor = hardwareMap.get(DistanceSensor.class, "rightDistanceSensor");
-//        backDistanceSensor = hardwareMap.get(DistanceSensor.class, "backDistanceSensor");
-
-        waitForStart();
-        // 4. Drive forward for 1 second
-        if (opModeIsActive()) {
-            odo.update();
-
-            // Update time variable
-            double newTime = getRuntime();
-            double loopTime = newTime - oldTime;
-            double frequency = 1 / loopTime;
-            oldTime = newTime;
-
-            // Get the position of the robot
-            Pose2D pos = odo.getPosition();
-            String data = String.format(Locale.US, "{X: %.3f, Y: %.3f, H: %.3f}", pos.getX(DistanceUnit.INCH), pos.getY(DistanceUnit.INCH), pos.getHeading(AngleUnit.DEGREES));
-            telemetry.addData("Position", data);
-
-            // Get velocity of the robot
-            String velocity = String.format(Locale.US,"{XVel: %.3f, YVel: %.3f, HVel: %.3f}", odo.getVelX(DistanceUnit.INCH), odo.getVelY(DistanceUnit.INCH), odo.getHeadingVelocity(UnnormalizedAngleUnit.DEGREES));
-            telemetry.addData("Velocity", velocity);
-
-            // Get status of the odometry (error stuff)
-            telemetry.addData("Status", odo.getDeviceStatus());
-            telemetry.addData("Pinpoint Frequency", odo.getFrequency()); //prints/gets the current refresh rate of the Pinpoint
-            telemetry.addData("REV Hub Frequency: ", frequency); //prints the control system refresh rate
-            telemetry.update();
-
-        }
-
-        //telemetry.addData("rmp flywheel: ", flywheelMotor.getPower());
-        //telemetry.addData("rotShot", rotShot.getPower());
-        //telemetry.update();
-
-
-    }
-}
